requirements.txt
transformers>=4.41.0
torch>=2.2.0
datasets>=2.19.0
pyyaml>=6.0.0
numpy>=1.26.0
tqdm>=4.66.0
pydantic>=2.7.0
# Required for 4-bit quantized checkpoint loading on CUDA
bitsandbytes>=0.43.0
accelerate>=0.30.0

# Development / testing
pytest>=8.0.0
pytest-cov>=5.0.0
